Minor security improvements.
