+++ app/jobs/routes.py
@@ -2,14 +2,14 @@ from flask import current_app, flash, redirect, render_template, url_for
 from flask_security import login_required
 from flask_login import current_user
 
-from app.jobs import bp, scheduler, controllers as controller
+from app.jobs import bp, controllers as controller
 from app.jobs.forms import ScheduleForm
 from tasks import tasks
 
 @bp.route('/jobs')
 @login_required
 def jobs_index():
-	return render_template("jobs/jobs.html", jobs=scheduler.get_user_jobs(current_user.get_id()))
+	return render_template("jobs/jobs.html", jobs=controller.get_user_jobs(current_user.get_id()))
 
 @bp.route('/schedule', methods=["GET", "POST"])
 @login_required
@@ -25,7 +25,7 @@ def schedule():
 			params.append(form[field].data)
 		print(params)
 		try:
-			scheduler.schedule_job(task, params, job_name)
+			controller.schedule_job(task, params, job_name)
 		except Exception as e:
 			if current_app.config["DEBUG"]:
 				raise e
@@ -41,7 +41,6 @@ def schedule():
 @bp.route('jobs/kill/<id>')
 @login_required
 def kill(id):
-	# TODO: Prevent users from cancelling each other's jobs
 	job = controller.get_job_entry(id)
 	if job is None:
 		flash("Job does not exist.")
@@ -53,8 +52,11 @@ def kill(id):
 @bp.route('jobs/view/<id>')
 @login_required
 def view(id):
-	# TODO: Prevent users from viewing each other's jobs?
 	job = controller.get_job_entry(id)
+	if job is None:
+		flash("Job does not exist.")
+		return redirect(url_for("jobs.jobs_index"))
+
 	results = controller.get_job_results(id)
 	if job.task == "word_freq":
 		return render_template("jobs/ngram_viewer.html", data = results)
